feat: complete main reconcile logic
