+++ .gitignore
@@ -1,4 +1,5 @@
 # Binaries for programs and plugins
+bin/
 *.exe
 *.exe~
 *.dll
@@ -15,3 +16,4 @@
 *vendor/
 *.DS_Store
 .idea
+.run
