+++ go.mod
@@ -5,6 +5,7 @@ go 1.16
 require (
 	github.com/go-logr/logr v0.3.0
 	github.com/go-test/deep v1.0.8
+	github.com/gogo/protobuf v1.3.1
 	github.com/kubernetes-csi/external-snapshotter/client/v4 v4.2.0
 	github.com/onsi/ginkgo v1.14.1
 	github.com/onsi/gomega v1.10.2
