+++ pkg/utils/compare.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License
+
+package utils
+
+func MaxFloat64(i float64, j ...float64) float64 {
+	maxValue := i
+	for _, j_ := range j {
+		if j_ > maxValue {
+			maxValue = j_
+		}
+	}
+	return maxValue
+}
+
+func MaxInt64(i, j int64) int64 {
+	if i > j {
+		return i
+	}
+	return j
+}
